requirements.txt
# Core
langchain>=0.3
langchain-core
langchain-community
langgraph>=0.3

# Retrieval / ML utils
faiss-cpu
rapidfuzz
httpx

# API & web
fastapi
uvicorn[standard]

# Data & DB
SQLAlchemy>=2.0
pydantic
python-dotenv
tabulate
pandas
scikit-learn
PyYAML

# Optional: durable checkpoints (Postgres)
psycopg2-binary

# Dev UX (optional but nice)
langgraph-cli
